various warnings fixes
git-svn-id: svn://svn.savannah.gnu.org/myserver/trunk@168 c4c922ea-a70d-0410-937d-ec1daf91936c

+++ myserverweb/include/cserver.h
@@ -81,7 +81,7 @@ private:
 	char path[MAX_PATH];
 	char browseDirCSSpath[MAX_PATH];
 	char *defaultFilename;
-	int nDefaultFilename;
+	u_long nDefaultFilename;
 	char serverAdmin[32];
 	ClientsTHREAD threads[MAXIMUM_PROCESSORS];
 	u_long nThreads;
